# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, file, plugin, mcp, skill. The config scope also
    grants bot and provider access.
servers:
  - url: http://localhost:6185
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/system-config:
    get:
      tags: [System Config]
      summary: Get the system configuration
      operationId: getSystemConfig
      x-astrbot-scope: config
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [System Config]
      summary: Replace the system configuration
      operationId: updateSystemConfig
      x-astrbot-scope: config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DynamicConfig"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /api/v1/config-profiles:
    get:
      tags: [Config Profiles]
      summary: List configuration profiles
      operationId: listConfigProfiles
      x-astrbot-scope: config
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Config Profiles]
      summary: Create a configuration profile
      operationId: createConfigProfile
      x-astrbot-scope: config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConfigProfileRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/config-profiles/{config_id}:
    get:
      tags: [Config Profiles]
      summary: Get a configuration profile
      operationId: getConfigProfile
      x-astrbot-scope: config
      parameters:
        - $ref: "#/components/parameters/ConfigId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Config Profiles]
      summary: Replace a configuration profile
      operationId: updateConfigProfileContent
      x-astrbot-scope: config
      parameters:
        - $ref: "#/components/parameters/ConfigId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DynamicConfig"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    patch:
      tags: [Config Profiles]
      summary: Rename a configuration profile
      operationId: renameConfigProfile
      x-astrbot-scope: config
      parameters:
        - $ref: "#/components/parameters/ConfigId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Config Profiles]
      summary: Delete a configuration profile
      operationId: deleteConfigProfile
      x-astrbot-scope: config
      parameters:
        - $ref: "#/components/parameters/ConfigId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/config-routes:
    get:
      tags: [Bot Config Routes]
      summary: List UMO to configuration profile bindings
      operationId: listConfigRoutes
      x-astrbot-scope: config
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Bot Config Routes]
      summary: Replace all UMO to configuration profile bindings
      operationId: replaceConfigRoutes
      x-astrbot-scope: config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigRoutesReplaceRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/config-routes/{umo}:
    put:
      tags: [Bot Config Routes]
      summary: Bind a UMO to a configuration profile
      operationId: upsertConfigRoute
      x-astrbot-scope: config
      parameters:
        - $ref: "#/components/parameters/Umo"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigRouteUpsertRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Bot Config Routes]
      summary: Remove a UMO configuration profile binding
      operationId: deleteConfigRoute
      x-astrbot-scope: config
      parameters:
        - $ref: "#/components/parameters/Umo"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /api/v1/chat/sessions/new:
    get:
      tags: [Chat]
      summary: Create a webchat session
      operationId: createChatSession
      x-astrbot-scope: chat
      parameters:
        - name: platform_id
          in: query
          schema:
            type: string
            default: webchat
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/sessions/batch-delete:
    post:
      tags: [Chat]
      summary: Delete multiple webchat sessions
      operationId: batchDeleteChatSessions
      x-astrbot-scope: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionBatchDeleteRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/sessions/{session_id}:
    get:
      tags: [Chat]
      summary: Get a webchat session and its messages
      operationId: getChatSession
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    patch:
      tags: [Chat]
      summary: Update a webchat session
      operationId: updateChatSession
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionPatchRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Chat]
      summary: Delete a webchat session
      operationId: deleteChatSession
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/sessions/{session_id}/stop:
    post:
      tags: [Chat]
      summary: Stop active webchat work for a session
      operationId: stopChatSession
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/sessions/{session_id}/messages/{message_id}:
    patch:
      tags: [Chat]
      summary: Update the latest user message in a webchat session
      operationId: updateChatMessage
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: message_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessagePatchRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/sessions/{session_id}/messages/{message_id}/regenerate:
    post:
      tags: [Chat]
      summary: Regenerate a bot message in a webchat session
      operationId: regenerateChatMessage
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: message_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageRegenerateRequest"
      responses:
        "200":
          description: Server-sent chat stream or an error envelope
...
  /api/v1/chat/threads:
    post:
      tags: [Chat]
      summary: Create or reuse a side thread from a webchat message
      operationId: createChatThread
      x-astrbot-scope: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatThreadCreateRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/threads/{thread_id}:
    get:
      tags: [Chat]
      summary: Get a webchat side thread
      operationId: getChatThread
      x-astrbot-scope: chat
      parameters:
        - name: thread_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Chat]
      summary: Delete a webchat side thread
      operationId: deleteChatThread
      x-astrbot-scope: chat
      parameters:
        - name: thread_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/threads/{thread_id}/messages:
    post:
      tags: [Chat]
      summary: Send a message inside a webchat side thread
      operationId: sendChatThreadMessage
      x-astrbot-scope: chat
      parameters:
        - name: thread_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatThreadMessageRequest"
      responses:
        "200":
          description: Server-sent chat stream or an error envelope
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
  parameters:
...
    ConfigId:
      name: config_id
      in: path
      required: true
      schema:
        type: string
...
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
...
    Umo:
      name: umo
      in: path
      required: true
      description: URL-encoded unified message origin.
      schema:
        type: string
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: string
        data: {}
      additionalProperties: false
...
    DynamicConfig:
      type: object
      additionalProperties: true
...
    CreateConfigProfileRequest:
      type: object
      properties:
        name:
          type: string
        config:
          $ref: "#/components/schemas/DynamicConfig"
      additionalProperties: false
    RenameRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
      additionalProperties: false
...
    ConfigRoutesReplaceRequest:
      type: object
      required: [routing]
      properties:
        routing:
          type: object
          additionalProperties:
            type: string
      additionalProperties: false
    ConfigRouteUpsertRequest:
      type: object
      required: [config_id]
      properties:
        config_id:
          type: string
          description: Use "default" to remove a custom route and fall back.
      additionalProperties: false
...
    ChatSessionBatchDeleteRequest:
      type: object
      required: [session_ids]
      properties:
        session_ids:
          type: array
          items:
            type: string
      additionalProperties: false
    ChatSessionPatchRequest:
      type: object
      properties:
        display_name:
          type: string
      additionalProperties: false
    ChatMessagePatchRequest:
      type: object
      required: [content]
      properties:
        content:
          type: object
          additionalProperties: true
      additionalProperties: false
    ChatMessageRegenerateRequest:
      type: object
      properties:
        selected_provider:
          type: string
        selected_model:
          type: string
        enable_streaming:
          type: boolean
          default: true
      additionalProperties: false
    ChatThreadCreateRequest:
      type: object
      required: [session_id, parent_message_id, selected_text]
      properties:
        session_id:
          type: string
        parent_message_id:
          oneOf:
            - type: string
            - type: integer
        selected_text:
          type: string
      additionalProperties: false
    ChatThreadMessageRequest:
      type: object
      required: [message]
      properties:
        message:
          oneOf:
            - type: string
            - type: array
              items:
                $ref: "#/components/schemas/MessagePart"
        selected_provider:
          type: string
        selected_model:
          type: string
        enable_streaming:
          type: boolean
          default: true
      additionalProperties: false
...
    MessagePart:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [text, plain, image, file, audio, record, video, reply]
        text:
          type: string
        attachment_id:
          type: string
        url:
          type: string
        filename:
          type: string
        mime_type:
          type: string
      additionalProperties: true
...